+++ interview/views.py
@@ -8,6 +8,7 @@ import json
 
 from django.contrib.admin.models import LogEntry, ADDITION, CHANGE
 from django.contrib.admin.options import get_content_type_for_model
+from django.contrib.auth.views import redirect_to_login
 from plotly.offline import plot
 import plotly.figure_factory as ff
 import plotly.express as px
@@ -241,6 +242,7 @@ def process(request, process_id, slug_info=None):
 
 @login_required
 @require_http_methods(["POST"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def close_process(request, process_id):
     try:
         process = Process.objects.for_user(request.user).get(pk=process_id)
@@ -258,6 +260,7 @@ def close_process(request, process_id):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def reopen_process(request, process_id):
     try:
         process = Process.objects.for_user(request.user).get(pk=process_id)
@@ -294,6 +297,9 @@ def closed_processes(request):
 @login_required
 @require_http_methods(["GET"])
 def processes_for_source(request, source_id):
+    if request.user.consultant.is_external and request.user.consultant.limited_to_source.id != source_id:
+        return redirect_to_login(next=request.path)
+
     try:
         source = Sources.objects.get(id=source_id)
     except Sources.DoesNotExist:
@@ -316,6 +322,7 @@ def processes_for_source(request, source_id):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def processes_for_offer(request, offer_id):
     try:
         offer = Offer.objects.get(id=offer_id)
@@ -400,6 +407,8 @@ def new_candidate(request, past_candidate_id=None):
                 process = process_form.save(commit=False)
                 process.candidate = candidate
                 process.creator = Consultant.objects.get(user=request.user)
+                if request.user.consultant.limited_to_source:
+                    process.sources = request.user.consultant.limited_to_source
                 process.save()
                 log_action(True, process, request.user, new_candidate)
 
@@ -416,6 +425,14 @@ def new_candidate(request, past_candidate_id=None):
         process_form = ProcessForm()
         interviewers_form = InterviewersForm(prefix="interviewers")
         process_form.fields["subsidiary"].initial = request.user.consultant.company.id
+
+    if request.user.consultant.is_external:
+        process_form.fields.pop("sources")
+
+        # restrict available interviewers to process creator
+        interviewers_form.fields["interviewers"].widget.queryset = interviewers_form.fields[
+            "interviewers"
+        ].queryset.filter(user=request.user)
 
     source_form = SourceForm(prefix="source")
     offer_form = OfferForm(prefix="offer")
@@ -463,13 +480,23 @@ def interview(request, process_id=None, interview_id=None, action=None):
         if action == "planning-request":
             interview_model.toggle_planning_request()
             return ret
+
+        if request.user.consultant.is_external:
+            # set interviewer to be external consultant
+            tmp = request.POST.copy()
+            tmp["interviewers"] = request.user.consultant.id
+            request.POST = tmp
         form = InterviewForm(request.POST, instance=interview_model)
+
         if form.is_valid():
             form.save()
             log_action(False, interview_model, request.user, interview)
             return ret
     else:
         form = InterviewForm(instance=interview_model)
+
+    if request.user.consultant.is_external:
+        form.fields.pop("interviewers", None)  # interviewer will always be user
 
     return render(
         request,
@@ -528,6 +555,9 @@ def minute(request, interview_id, slug_info=None):
 @login_required
 @require_http_methods(["GET"])
 def dashboard(request):
+    if request.user.consultant.limited_to_source:  # if None, dashboard will be empty anyways
+        return processes_for_source(request, request.user.consultant.limited_to_source.id)
+
     a_week_ago = timezone.now() - datetime.timedelta(days=7)
     c = request.user.consultant
     actions_needed_processes = (
@@ -568,6 +598,7 @@ def dashboard(request):
 
 @login_required
 @require_http_methods(["POST"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def create_source_ajax(request):
     form = SourceForm(request.POST, prefix="source")
     if form.is_valid():
@@ -582,6 +613,7 @@ def create_source_ajax(request):
 
 @login_required
 @require_http_methods(["POST"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def create_offer_ajax(request):
     form = OfferForm(request.POST, prefix="offer")
     if form.is_valid():
@@ -597,6 +629,7 @@ def create_offer_ajax(request):
 @csrf_exempt
 @require_http_methods(["POST"])
 @user_passes_test(lambda u: u.is_superuser)
+@user_passes_test(lambda u: not u.consultant.is_external)
 def create_account(request):
     data = json.loads(request.body)
     subsidiary = Subsidiary.objects.filter(code=data["company"]).first()
@@ -623,6 +656,7 @@ def create_account(request):
 @csrf_exempt
 @require_http_methods(["DELETE"])
 @user_passes_test(lambda u: u.is_superuser)
+@user_passes_test(lambda u: not u.consultant.is_external)
 def delete_account(request, trigramme):
     user = PyouPyouUser.objects.filter(trigramme=trigramme.lower()).first()
     if not user:
@@ -663,6 +697,9 @@ def edit_candidate(request, process_id):
 
     source_form = SourceForm(prefix="source")
     offer_form = OfferForm(prefix="offer")
+
+    if request.user.consultant.is_external:
+        process_form.fields.pop("sources")
 
     data = {
         "process": process,
@@ -965,6 +1002,7 @@ def calculate_load(itws):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def interviewers_load(request, subsidiary_id=None):
     subsidiary = None
     if subsidiary_id:
@@ -1006,6 +1044,7 @@ def interviewers_load(request, subsidiary_id=None):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def search(request):
     q = request.GET.get("q", "")
 
@@ -1041,6 +1080,7 @@ RE_DESCRIPTION = re.compile(r"DESCRIPTION:(?P<description>(.|\n)*)LAST-MODIFIED"
 
 @login_required
 @require_http_methods(["GET", "POST"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def import_seekube(request):
     if request.method == "GET":
         form = UploadSeekubeFileForm()
@@ -1092,6 +1132,7 @@ def import_seekube(request):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def gantt(request):
     state_filter = Process.OPEN_STATE_VALUES + [Process.JOB_OFFER, Process.HIRED]
     today = timezone.now().date()
@@ -1206,6 +1247,7 @@ class OffersTable(tables.Table):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def active_sources(request, subsidiary_id=None):
     subsidiary = None
     sources_qs = Sources.objects.filter(archived=False)
@@ -1255,6 +1297,7 @@ def active_sources(request, subsidiary_id=None):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def offers(request, subsidiary_id=None):
     subsidiary = None
     offers_qs = Offer.objects.filter(archived=False)
@@ -1304,6 +1347,7 @@ def offers(request, subsidiary_id=None):
 
 @login_required
 @require_http_methods(["GET"])
+@user_passes_test(lambda u: not u.consultant.is_external)
 def activity_summary(request):
     process_filter = ProcessSummaryFilter(request.GET, queryset=Process.objects.all())
     interview_filter = InterviewSummaryFilter(request.GET, queryset=Interview.objects.all())