+++ pyoupyou/settings/common.py
@@ -47,6 +47,7 @@ MIDDLEWARE = [
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
+    "pyoupyou.middleware.ExternalCheckMiddleware",
 ]
 
 ROOT_URLCONF = "pyoupyou.urls"